Refactor: Media: componentize into smaller files
