+++ src/components/details/DetailsPanel.tsx
@@ -8,8 +8,8 @@ import { GlobalContext, LangOrEndoIntro, ScrollToTopOnMount } from 'components'
 import { LegendSwatch } from 'components/legend'
 import { RecordDescription } from 'components/results'
 import { paths as routes } from 'components/config/routes'
+import { Media } from 'components/media'
 import { useSymbAndLabelState } from '../../context/SymbAndLabelContext'
-import { Media } from './Media'
 import { useStyles } from './styles'
 import { LangRecordSchema } from '../../context/types'
